backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types
anyio 
attrs==25.3.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-sql-python-connector==1.18.1
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi
firebase-admin==6.8.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-speech==2.32.0
google-cloud-storage==3.1.0
google-cloud-texttospeech==2.27.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
httptools==0.6.4
idna 
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.4.3
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic 
pydantic_core 
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pytest==8.3.5
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
setuptools==75.8.0
sniffio 
SQLAlchemy==2.0.40
starlette 
tqdm==4.67.1
typing_extensions 
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
yarl==1.20.0
